Rewritten user validation in UserManager class, solution to pass data between ViewCotrollers without segue
UserManager class - rewritten  guards and if's which helps to check user input when register or login:
username and password must be at least 8 symbols, username must nor be registered, password and retype password must be the same.
Implemented temporary solution for passing values (username and moneyAmount) when navigating from RootVC to TransfersVC, after user was created or logged.